requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.6.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
decorator==5.1.1
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
flower==1.2.0
greenlet==2.0.1
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.12.0
incremental==22.10.0
ipdb==0.13.9
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
numpy==1.23.5
openpyxl==3.1.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyPBD==2.1.3
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.4.2
requests==2.28.2
service-identity==21.1.0
six==1.16.0
SQLAlchemy==2.0.0
sqlparse==0.4.3
stack-data==0.6.1
toml==0.10.2
tornado==6.2
traitlets==5.5.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
zope.interface==5.5.2
